requirements.txt
streamlit
nltk
scikit-learn
tweepy
python-dotenv
certifi